cleaning up creators code
