+++ packages/cli/src/tokens/create.ts
@@ -1,12 +1,10 @@
 import type { ColorScheme } from '../colors/types.js';
 import { getDefaultToken, getDefaultTokens } from './create/defaults.js';
-import {
-  generateColorScheme,
-  generateGlobal,
-  generateSemantic,
-  generateTheme,
-  generateTypography,
-} from './create/generators.js';
+import { generateColorGlobal, generateColorScheme } from './create/generators/color.js';
+import { generateSemantic } from './create/generators/semantic.js';
+import { generateTheme } from './create/generators/theme.js';
+import { generateTypography } from './create/generators/typography.js';
+
 import type { Theme, TokenSet, TokenSets } from './types.js';
 
 export const cliOptions = {
@@ -45,7 +43,7 @@ export const createTokens = async (opts: Theme) => {
     [`primitives/modes/typography/primary/${name}`, generateTypography(name, typography)],
     [`primitives/modes/typography/secondary/${name}`, generateTypography(name, typography)],
     ...colorSchemes.flatMap((scheme): [string, TokenSet][] => [
-      [`primitives/modes/color-scheme/${scheme}/global`, generateGlobal(scheme)],
+      [`primitives/modes/color-scheme/${scheme}/global`, generateColorGlobal(scheme)],
       [`primitives/modes/color-scheme/${scheme}/${name}`, generateColorScheme(name, scheme, colors)],
     ]),
     [`themes/${name}`, generateTheme(colors, name, borderRadius)],
