+++ packages/cli/src/tokens/create/generators/theme.ts
@@ -1,0 +1,48 @@
+import * as R from 'ramda';
+import themeBase from '../../template/design-tokens/themes/theme.base.template.json' with { type: 'json' };
+import themeColorTemplate from '../../template/design-tokens/themes/theme.template.json' with { type: 'json' };
+
+import type { Colors, TokenSet } from '../../types.js';
+
+export const generateTheme = (colors: Colors, themeName: string, borderRadius: number) => {
+  const mainColorNames = Object.keys(colors.main);
+  const supportColorNames = Object.keys(colors.support);
+  const customColors = [...mainColorNames, 'neutral', ...supportColorNames];
+
+  const themeColorTokens = Object.fromEntries(
+    customColors.map(
+      (colorName) =>
+        [
+          colorName,
+          R.map((x) => ({ ...x, $value: x.$value.replace('<color>', colorName) }), themeColorTemplate),
+        ] as const,
+    ),
+  );
+
+  const { color: themeBaseFileColor, ...remainingThemeFile } = themeBase;
+  const themeFile = {
+    color: {
+      ...themeColorTokens,
+      ...themeBaseFileColor,
+    },
+    ...remainingThemeFile,
+  };
+
+  const baseBorderRadius = R.lensPath(['border-radius', 'base', '$value']);
+  const updatedThemeFile = R.set(baseBorderRadius, String(borderRadius), themeFile);
+  const token = JSON.parse(
+    JSON.stringify(
+      updatedThemeFile,
+      (key, value) => {
+        if (key === '$value') {
+          return (value as string).replace('<theme>', themeName);
+        }
+
+        return value;
+      },
+      2,
+    ),
+  ) as TokenSet;
+
+  return token;
+};
